Update held gizmos during update. Disable vsync
